prediction_project/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.1
plotly==5.18.0
pandas==2.0.3
scikit-learn==1.3.2
django-allauth==0.61.0
whitenoise==6.6.0
gunicorn==21.2.0
